+++ app/Http/Controllers/Admin/Car/Transmission/TransmissionController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Admin\ResponseTrait;
 use App\Http\Controllers\Controller;
 use App\Models\Transmission;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Yajra\DataTables\Facades\DataTables;
 
 class TransmissionController extends Controller
@@ -14,12 +15,13 @@ class TransmissionController extends Controller
 
     public function index(Request $request)
     {
-
+        Gate::authorize('transmission.view');
         return view('admin.pages.transmission.index');
     }
 
     public function store(Request $request)
     {
+        Gate::authorize('transmission.create');
         $rules = [];
         foreach (locales() as $key => $language) {
             $rules['name_' . $key] = 'required|string|max:255';
@@ -36,6 +38,7 @@ class TransmissionController extends Controller
 
     public function update(Request $request)
     {
+        Gate::authorize('transmission.update');
         $rules = [];
         foreach (locales() as $key => $language) {
             $rules['name_' . $key] = 'required|string|max:255';
@@ -52,6 +55,7 @@ class TransmissionController extends Controller
 
     public function destroy($uuid)
     {
+        Gate::authorize('transmission.delete');
         $uuids=explode(',', $uuid);
         Transmission::whereIn('uuid', $uuids)->delete();
         return $this->sendResponse(null, null);
@@ -100,6 +104,7 @@ class TransmissionController extends Controller
 
     public function activate($uuid)
     {
+        Gate::authorize('transmission.update');
         $activate =  Transmission::findOrFail($uuid);
         $activate->status = !$activate->status;
         if (isset($activate) && $activate->save()) {
