+++ app/Http/Controllers/Admin/places/country/CountryController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Admin\ResponseTrait;
 use App\Http\Controllers\Controller;
 use App\Models\Country;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Yajra\DataTables\Facades\DataTables;
 
 class CountryController extends Controller
@@ -14,12 +15,13 @@ class CountryController extends Controller
 
     public function index(Request $request)
     {
-
+        Gate::authorize('place.view');
         return view('admin.pages.country.index');
     }
 
     public function store(Request $request)
     {
+        Gate::authorize('place.create');
         $rules = [];
         foreach (locales() as $key => $language) {
             $rules['name_' . $key] = 'required|string|max:255';
@@ -37,6 +39,7 @@ class CountryController extends Controller
 
     public function update(Request $request)
     {
+        Gate::authorize('place.update');
         $rules = [];
         foreach (locales() as $key => $language) {
             $rules['name_' . $key] = 'required|string|max:255';
@@ -53,6 +56,7 @@ class CountryController extends Controller
 
     public function destroy($uuid)
     {
+        Gate::authorize('place.delete');
         $uuids=explode(',', $uuid);
         Country::whereIn('uuid', $uuids)->delete();
         return $this->sendResponse(null, null);
@@ -102,6 +106,7 @@ class CountryController extends Controller
 
     public function activate($uuid)
     {
+        Gate::authorize('place.update');
         $activate =  Country::findOrFail($uuid);
         $activate->status = !$activate->status;
         if (isset($activate) && $activate->save()) {
